requirements.txt
# flask
flask
# mysql
mysqlclient
mysql-connector
sqlalchemy
pymysql
MYSQL-connector-python
flask_mysqldb
# util
numpy
pandas
pandasql
tqdm
python-dotenv
opencv-python